# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "eCommerce",
    "description": "Documentation for eCommerce Exam Project"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/cart/": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "Get logged in user cart",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "Create new category",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AddToCart"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Cart"
        ],
        "description": "Update Cart product quantity",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCartProduct"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "description": "Delete Cart product.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/DeleteCartProduct"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/cart/checkout": {
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "Checkout logged in user cart",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Get all categories in database",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Create new category. Can be anything except for a number",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateCategory"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Update category",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCategory"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Delete category based on category name",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/DeleteCategory"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/admin/categories": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get categories page for admin with crud functionality",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateCategory": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "SwaggerCategoryTest"
        }
      },
      "required": [
        "category"
      ]
    },
    "UpdateCategory": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        },
        "updatedCategory": {
          "type": "string",
          "example": "NewCategoryName"
        }
      },
      "required": [
        "id",
        "updatedCategory"
      ]
    },
    "DeleteCategory": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "SwaggerCategoryTest"
        }
      },
      "required": [
        "category"
      ]
    },
...
    "AddToCart": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string",
          "example": "1"
        },
        "quantity": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "productId",
        "quantity"
      ]
    },
    "UpdateCartProduct": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string",
          "example": "1"
        },
        "quantity": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "productId",
        "quantity"
      ]
    },
    "DeleteCartProduct": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "productId"
      ]
    },
...
  }
}